+++ src/control_plane/migration/postgres/mod.rs
@@ -3,6 +3,10 @@ mod tests;
 
 mod backup_postgres_database;
 mod postgres_backup_options;
+mod postgres_restore_options;
+mod restore_postgres_database;
 
 pub(crate) use backup_postgres_database::backup_postgres_database;
 pub(crate) use postgres_backup_options::PostgresBackupOptions;
+pub(crate) use postgres_restore_options::PostgresRestoreOptions;
+pub(crate) use restore_postgres_database::restore_postgres_database;